Minor bug fix to kleeReadMemory
We forgot to read the third parameter and were just ignoring it.  This
new implementation does not add a constraint when concretizing I/O
memory--the old bugged version did add the constraint.  To add
constraints permanently, it's necessary to change
target-i386/op_helper.c and pass "true" for the third parameter to
tcg_llvm_get_value instead of "false."